.history/requirements_20260401165410.txt
streamlit 
scikit-learn 
rapidfuzz
pytesseract 
pdf2image 
PyPDF2 
pillow